% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial}
\alias{factorial}
\title{Compute factorial}
\usage{
factorial(n)
}
\arguments{
\item{n}{argument to factorial function}
}
\value{
factorial of \code{n}
}
\description{
Compute factorial
}
\examples{
factorial(0)
factorial(1)
factorial(4)
}
