% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl.pkgdown.demo.ghp-package.R
\docType{package}
\name{hl.pkgdown.demo.ghp-package}
\alias{hl.pkgdown.demo.ghp}
\alias{hl.pkgdown.demo.ghp-package}
\title{hl.pkgdown.demo.ghp: My miscellaneous utilities, documented with {pkgdown}}
\description{
This package is right now just a sandbox to study what {pkgdown} (and GitHub Actions) can do for me (and maybe you).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HenningLorenzen-ext-bayer/pkgdown-demo-ghp}
  \item \url{https://henninglorenzen-ext-bayer.github.io/pkgdown-demo-ghp/}
  \item Report bugs at \url{https://github.com/HenningLorenzen-ext-bayer/pkgdown-demo-ghp/issues}
}

}
\keyword{internal}
